drm/sde/rotator: Remove unneeded PM QoS requests.
When the rotator is actually used (still an unsolved question in
computer science), these PM QoS requests block some CPUs in the LITTLE
cluster from entering deep idle because the driver assumes that display
rotating work occurs on a hardcoded set of CPUs, which is false. We
already have the IRQ PM QoS machinery for display rendering operations
that actually matter, so this cruft is unneeded.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Vishalcj17 <vishalcj@aospa.co>
Change-Id: I01e03e512ef9b37fc911c623594588a73cdcc4fb
Signed-off-by: mcdofrenchfreis <xyzevan@androidist.net>