Revert "mm: reduce the time spend by killed tasks in alloc path"
This reverts commit ac80551df4aa65c46123da65bd9f9676cb0b2079.

Returning NULL from the page allocator just because the current process
is dying produces dangerous consequences in places where page allocation
failure is either unexpected or treated as a bona fide indicator that
memory has run out. In general, for non-costly page allocations which
lack the __GFP_NORETRY flag, every effort should be made to satisfy the
allocation in order to avoid the aforementioned consequences. This fatal
signal check would be fine to keep if just about every allocation site
were aware of the fatal signal being the cause of the allocation failure,
but unfortunately that is not the case.

Remove the fatal signal optimization since it changes the semantics of
the page allocator in a way that is unexpected by kernel code.

[Kazuki: Use Sultan's commit message]
Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>
Signed-off-by: mcdofrenchfreis <xyzevan@androidist.net>