thaw_fingerprintd: Fix strstr() usage
To improve the accuracy of fingerprint recognition processes, the "thaw_fingerprintd"
module's code was adjusted. The initial use of the "strstr()" function had an error in
its condition, leading to wrong detection. The recent update fixes this by properly using
"strstr()" to check for the term "fingerprint" in the process name.

Signed-off-by: mcdofrenchfreis <xyzevan@androidist.net>